fix(data): handle fractional Nightscout threshold values
Same failure mode as sgv: ThresholdsDto declared bgHigh/bgTargetTop/
bgTargetBottom/bgLow as Int, so a fractional value in Nightscout's
settings.thresholds would crash JSON parsing. Parse them as Double
and round to the nearest mg/dL when building GlucoseThresholds —
thresholds are only used for alert comparisons and graph pixel
placement, never displayed with sub-integer precision, so rounding
at the domain boundary is sufficient for both mg/dL and mmol/L.

Refs #13

+++ app/src/main/java/de/autosugar/data/network/dto/StatusDto.kt
@@ -22,8 +22,8 @@ data class SettingsDto(
 
 @JsonClass(generateAdapter = true)
 data class ThresholdsDto(
-    val bgHigh: Int?,
-    val bgTargetTop: Int?,
-    val bgTargetBottom: Int?,
-    val bgLow: Int?,
+    val bgHigh: Double?,
+    val bgTargetTop: Double?,
+    val bgTargetBottom: Double?,
+    val bgLow: Double?,
 )
